CNAME?
This site is served from an Apex domain. GitHub seemed to add this file on my behalf when flipping settings in the settings -> pages section.

It may be that we can't actually get both xor.lonnen.com and lonnen.com domains both from GitHub pages. For a long time I wasn't using lonnen.com at all. If one domain must win, it will also serve as the root of other GH Pages-hosted content. Rather than muddle those projects with the blog posts of XOR it may be better to use the single-page static website as the base so there won't be conflicts